+++ pi-extension/src/core/agents-config/files-config.ts
@@ -1,7 +1,7 @@
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { looksLikeTestPath } from "./files-discovery.js";
-import { atomicWriteJson } from "../io/atomic-write.js";
+import { looksLikeTestPath } from "./files-types.js";
+import { atomicWriteJson } from "../../io/atomic-write.js";
 
 export const FILES_CONFIG_FILE = "files.json";